// server/auth-service/src/interfaces/IModel/IUser.ts
import { Document } from 'mongoose';
import { UserRole } from '../../types';

export interface IUser extends Document {
    id: string; // From transform function
    fullName: string;
    email: string;
    phone: string;
    password?: string;
    role: UserRole;
    tenantId?: string;
    isActive: boolean;
    isVerified: boolean;
    emailVerifiedAt?: Date;
    phoneVerifiedAt?: Date;
    profilePicture?: string;
    lastLoginAt?: Date;
    googleId?: string;
    createdAt: Date;
    updatedAt: Date;
}